backend/requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
Pillow